refactor: BettingResult를 Betting 패키지로 이동
