+++ src/main/java/domain/betting/policy/BettingPolicy.java
@@ -1,6 +1,6 @@
 package domain.betting.policy;
 
-import domain.analyzer.BettingResult;
+import domain.betting.BettingResult;
 import domain.betting.BettingRate;
 import domain.gamer.Dealer;
 import domain.gamer.Player;
